+++ ansiplus/tests/testinput.py
@@ -1,5 +1,7 @@
 
 if __name__ == "__main__":
-    import ansiplus
-    mytext = ansiplus.input_color("text prompt example (green): ", "green")
+    from ansiplus import input_color
+    mytext = input_color("text prompt example (green): ", "green")
+    print("your text:", mytext)
+    mytext = input_color("colored prompt test (red): ", prompt_color="red")
     print("your text:", mytext)